Updates exponential distribution function and changes distance to cosine
