+++ compare_gridsearch.py
@@ -7,7 +7,6 @@ import pandas as pd
 import os
 import sys
 import matplotlib.pyplot as plt
-from scipy.optimize import curve_fit
 from sklearn.cluster import HDBSCAN
 from sklearn import utils
 from collections import defaultdict
@@ -16,11 +15,7 @@ import seaborn as sns
 import random
 from multiprocessing import Pool
 from functools import partial
-
-try:  # sklearn >= 0.22
-    from sklearn.neighbors import KernelDensity
-except ImportError:
-    from sklearn.neighbors.kde import KernelDensity
+from run_elfi import negative_exponential, fit_negative_exponential, plot_scatter
 
 def get_options():
     description = 'Compare parameters from a gridsearch of Pansim runs'
@@ -44,54 +39,6 @@ def get_options():
                     help='Number of threads. Default = 1.')            
 
     return parser.parse_args()
-
-# Copyright John Lees and Nicholas Croucher 2025
-def get_grid(minimum, maximum, resolution):
-    x = np.linspace(minimum, maximum, resolution)
-    y = np.linspace(minimum, maximum, resolution)
-    xx, yy = np.meshgrid(x, y)
-    xy = np.vstack([yy.ravel(), xx.ravel()]).T
-
-    return(xx, yy, xy)
-
-# Copyright John Lees and Nicholas Croucher 2025
-def plot_scatter(X, out_prefix, x_fit, y_fit):
-    # Plot results - max 1M for speed
-    max_plot_samples = 1000000
-    if X.shape[0] > max_plot_samples:
-        X = utils.shuffle(X, random_state=random.randint(1,10000))[0:max_plot_samples,]
-
-    # Kernel estimate uses scaled data 0-1 on each axis
-    scale = np.amax(X, axis = 0)
-    X /= scale
-
-    plt.figure(figsize=(11, 8), dpi= 160, facecolor='w', edgecolor='k')
-    xx, yy, xy = get_grid(0, 1, 100)
-
-    # KDE estimate
-    kde = KernelDensity(bandwidth=0.03, metric='euclidean',
-                        kernel='epanechnikov', algorithm='ball_tree')
-    kde.fit(X)
-    z = np.exp(kde.score_samples(xy))
-    z = z.reshape(xx.shape).T
-
-    levels = np.linspace(z.min(), z.max(), 10)
-    # Rescale contours
-    plt.contour(xx*scale[0], yy*scale[1], z, levels=levels[1:], cmap='plasma')
-    scatter_alpha = 1
-
-    # Plot on correct scale
-    plt.scatter(X[:,0]*scale[0].flat, X[:,1]*scale[1].flat, s=1, alpha=scatter_alpha)
-    plt.plot(x_fit, y_fit, label=f"Negative exponential 3 param", color='red')
-
-    plt.xlabel('Core distance (' + r'$\pi$' + ')')
-    plt.ylabel('Accessory distance (' + r'$a$' + ')')
-    plt.savefig(out_prefix + '_contours.png')
-    plt.close()
-
-# fit asymptotic curve using exponential decay
-def negative_exponential(x, b0, b1, b2): # based on https://isem-cueb-ztian.github.io/Intro-Econometrics-2017/handouts/lecture_notes/lecture_10/lecture_10.pdf and https://www.statforbiology.com/articles/usefulequations/
-    return b0 * (1 - np.exp(-b1 * x)) + b2
 
 def generate_summary_stats(file, plot, outpref):
     filename = os.path.splitext(os.path.basename(file))[0]
@@ -162,7 +109,7 @@ def generate_summary_stats(file, plot, outpref):
 
     # try negative_exponential model fit
     try:
-        popt, pcov = curve_fit(negative_exponential, obs_df[:,0], obs_df[:,1], p0=[1.0, 1.0, 0.0], bounds=([0.0, 0.0, 0.0], [1.0, np.inf, 1.0]))
+        popt, pcov = fit_negative_exponential(obs_df[:,0], obs_df[:,1])
         b0, b1, b2 = popt
         # get 1 std deviation error of parameters
         b0_err, b1_err, b2_err = np.sqrt(np.diag(pcov))
@@ -170,9 +117,9 @@ def generate_summary_stats(file, plot, outpref):
         values_dict["b0"] = b0
         values_dict["b1"] = b1
         values_dict["b2"] = b2
-        values_dict["b0_err"] = b0_err
-        values_dict["b1_err"] = b1_err
-        values_dict["b2_err"] = b2_err
+        values_dict["b0_err"] = b0_err / b0
+        values_dict["b1_err"] = b1_err / b1
+        values_dict["b2_err"] = b2_err / b2
 
         x_fit = np.linspace(0, obs_df[:,0].max(), 100)
         y_fit = negative_exponential(x_fit, *popt)
